Merge

+++ scripts/translate_to_rdf_alt_worker.sh
@@ -58,6 +58,10 @@ do
 
    java -jar $SAXON -s:$pdbml_valid_file -xsl:$PDBMLV2RDF_XSL -o:$rdf_valid_file 2> $err_file || ( cat $err_file && exit 1 )
 
+   if [ $has_rapper_command != "false" ] ; then
+    rapper -q -c $rdf_valid_file 2> $err_file || ( cat $err_file && exit 1 )
+   fi
+
    rm -f $err_file
 
    if [ $proc_id_mod = 0 ] ; then
